No longer copy profile name when copying session
Create-quantum-job was throwing "botocore.exceptions.ProfileNotFound:
The config profile (default) could not be found" since the container
does not have a profile "default" configuration. There are some
potential workarounds but the easiest is to just remove setting
the profile name all together. Since we are already copying the
credentials and region there really is no value in copying the profile
hence this change.