More integ test bucket creation exception handling (#100)
If a developer has already run the integ tests in one region and then
tries to run them from another, the tests will fail with an
IllegalLocationConstraintException, which is the error thrown when a
user attempts to access an existing bucket from _outside_ the region it
lives. This commit adds IllegalLocationConstraintException to the
exceptions to ignore.