Start polling for result of a task only when the customer wants it (#109)
Start polling for result of a task only when the customer wants it

Today, the SDK starts polling for task results as soon as the task object is created.
However, we can delay the polling until the customer actually tries to retrieve the result.
There is no impact on result retrieval latency due to this change. We are only postponing
the polling process until it's absolutely necessary.