fix: Always check region in AwsDevice.get_devices
AwsDevice.get_devices() creates a new AwsSession object for each region
that devices can be found and then creates an AwsDevice object for each
device found, using the last AwsSession created. However, there is no
guarantee that each device even exists in the session's region.

This is fine for QPUs, since the correct region is automatically used;
this is not the case for simulators, which have been assumed to exist in
every region. TN1 does not exist in every region, get_devices() will
always fail.

This commit fixes the bug by only instantiating simulators that are
available in the current region in AwsDevice.get_devices.

Also added integ tests for TN1.