fix: Get regions for QPUs instead of providers
Overhauls get_devices logic to filter regions based on ARNs instead of
providers. This future-proofs the SDK for QPUs added to new regions.

By getting rid of the region lists, this also fixes a bug where the
wrong devices would be returned for regions other than us-west-2, which
was the first region in the list.