change: update result_types to use hashing
use a dict to hold the result types to speed up checking if a result type is in _result_types. result_types property still interfaces as a list. Python 3.7+ officially guarantees order for dicts, so result_types is expected to be in order

also edit gate_model_quantum_task_result to add a dict mapping result type to index, as well as ensuring consistency regarding result type in the implementation