feature: Calculate arbitrary observables when `shots=0` (#267)
Allows arbitrary observables to be attached to a circuit, allowing, for example, the simultaneous calculation of all terms of a qubit Hamiltonian when `shots=0`. However, running such a circuit with `shots>0`, will throw an exception, as observables must commute to be measured (sampled) simultaneously; in fact, here the observables must be either identical or the identity.