fix: Validate hybrid jobs name (#1098)
Feature: Validation of job name when creating a hybrid job. 

Details: Implements a check for the validity of the quantum job name before attempting to create the job (resulting in a more descriptive error message than before). The check is in the job creation module, so it is executed regardless of the type of the job (Local or Aws) or its mode of creation (decorator or script). The doc-strings for job names in the different modules were updated for consistency.